Move pause into Agent.scala
Makes it easy to use LazyLogging.
