+++ src/main/scala/dispatch/github/package.scala
@@ -21,20 +21,4 @@ package object github {
       case str => nextPageRegex.findFirstIn(str)
     }
 
-  private[github] def pause(resp : Response)(implicit ec : ExecutionContext) 
-                           : Future[Boolean] = Future {
-    val hdrs = resp.getHeaders()
-    hdrs.getFirstValue("X-RateLimit-Remaining") match {
-      case "0" => {
-        val reset = 1000 * hdrs.getFirstValue("X-RateLimit-Reset").toLong
-        val now = System.currentTimeMillis()
-        // An extra second for clock skew.
-        val delay = math.max(1000, 1000 + reset - now)
-        Thread.sleep(delay)
-        true
-      }
-      case _ => false
-    }
-  }
-
 }
